% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Function to plot species richness or abundance}
\usage{
plot_abundance(data = dat, type = "abundance", interval = 3, col = "black")
}
\arguments{
\item{data}{Object with data in. Must have columns called Species and Time.}

\item{type}{Either abundance or richness.}

\item{interval}{Time interval (in hours) to group observations in.}

\item{col}{Colour to plot bars in.}
}
\description{
This function plots total abundance or species richness in time bins.
}
\author{
Martin Sullivan
}
