% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species.R
\name{plot_species}
\alias{plot_species}
\title{Function to plot species observations against time}
\usage{
plot_species(data = dat, col = "black")
}
\arguments{
\item{data}{Data object. Must have columns "Species" and "Time".}

\item{col}{Colour to plot points in.}
}
\description{
This function plots when each species was observed.
}
\author{
Martin Sullivan
}
