% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Function to read and format camera trap data}
\usage{
read_data()
}
\description{
This function let's the user select the file to read in, then checks column names and formats times so they are numeric. If names in data don't match expected ones, there are interactive prompts to select the one with time and species data.
}
\author{
Martin Sullivan
}
